refactor: activity ticker sub components
